refactor: rename user package to auth and add model and repository packages
